refactor license
